+++ admin/server/api/item/update.js
@@ -8,7 +8,6 @@ module.exports = function (req, res) {
 	const { params: { id }, body, locales } = req;
 	// locales, list: { options: { multilingual }, fields } } = req;
 	const newData = req.list.prepareCorrectParam(body);
-
 	req.list.model.findById(id, function (err, item) {
 		if (err) return res.status(500).json({ error: req.t.__('msg_db_error_withoutReason'), detail: err });
 		if (!item) return res.status(404).json({ error: req.t.__('msg_user_notfound'), id });
@@ -23,6 +22,7 @@ module.exports = function (req, res) {
 			frontLang: locales && locales.frontLang,
 		};
 		req.list.updateItem(item, newData, options, function (err) {
+			console.log(err);
 			if (err) {
 				var status = err.error === 'validation errors' ? 400 : 500;
 				const error = err.error === 'database error' ? req.t.__('msg_db_error_without', {